+++ Editor/Comparer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4a42e6bf5f4e3d4cb9fdda50d429a13
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
