+++ Editor/Comparer/ShaderVarintInfoComparer.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4169a7633ef3e764489115ad7b2eab95
