+++ Editor/ShaderVariantStripLogger.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 72f3da003dd981d418304c2d2a418386
