+++ Editor/UI/StrippingConfigForUnity6Window.cs.meta
