+++ Editor/UI/StrippingConfigWindow.cs
