+++ Editor/UI/StrippingConfigWindow.cs.meta
